Refactored SSL code
Separated SSL methods from the socket library into separate ssl library and made the changes to along with it including the changes to the tests.  Now the responses are stored in the logs.
