+++ src/nettacker/core/utils/common.py
@@ -7,8 +7,6 @@ import math
 import multiprocessing
 import random
 import re
-import socket
-import ssl
 import string
 import sys
 import time
@@ -326,98 +324,3 @@ def sort_dictionary(dictionary):
     if etc_flag:
         sorted_dictionary["..."] = {}
     return sorted_dictionary
-
-
-def is_weak_hash_algo(algo):
-    algo = algo.lower()
-    for unsafe_algo in ("md2", "md4", "md5", "sha1"):
-        if unsafe_algo in algo:
-            return True
-    return False
-
-
-def is_weak_ssl_version(host, port, timeout):
-    def test_ssl_verison(host, port, timeout, ssl_version=None):
-        try:
-            context = ssl.SSLContext(ssl_version)
-            socket_connection = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-            socket_connection.settimeout(timeout)
-            socket_connection.connect((host, port))
-            socket_connection = context.wrap_socket(socket_connection, server_hostname=host)
-            return socket_connection.version()
-
-        except ssl.SSLError:
-            return False
-
-        except (socket.timeout, ConnectionRefusedError, ConnectionResetError):
-            return None
-
-    ssl_versions = (
-        ssl.PROTOCOL_TLS_CLIENT,  # TLS 1.3
-        ssl.PROTOCOL_TLSv1_2,
-        ssl.PROTOCOL_TLSv1_1,
-        ssl.PROTOCOL_TLSv1,
-    )
-    lowest_version = None
-
-    for ssl_version in ssl_versions:
-        version = test_ssl_verison(host, port, timeout, ssl_version=ssl_version)
-        if version:
-            lowest_version = version
-
-    return (lowest_version, lowest_version not in {"TLSv1.2", "TLSv1.3"})
-
-
-def is_weak_cipher_suite(host, port, timeout):
-    def test_single_cipher(host, port, cipher, timeout):
-        try:
-            context = ssl.create_default_context()
-            context.check_hostname = False
-            context.verify_mode = ssl.CERT_NONE
-            context.set_ciphers(cipher)
-
-            socket_connection = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-            socket_connection.settimeout(timeout)
-            socket_connection.connect((host, port))
-            socket_connection = context.wrap_socket(socket_connection, server_hostname=host)
-            return True
-
-        except ssl.SSLError:
-            return False
-
-        except (socket.timeout, ConnectionRefusedError, ConnectionResetError):
-            return None
-
-    cipher_suites = [
-        "HIGH",  # OpenSSL cipher strings
-        "MEDIUM",
-        "LOW",
-        "EXP",
-        "eNULL",
-        "aNULL",
-        "RC4",
-        "DES",
-        "MD5",
-        "SHA1",
-        "DH",
-        "ADH",
-        "DHE",
-        "ECDH",
-        "ECDHE",
-        "TLSv1",
-        "TLSv1.1",
-        "TLSv1.2",
-        "TLSv1.3",
-    ]
-
-    supported_ciphers = []
-    for cipher in cipher_suites:
-        if test_single_cipher(host, port, cipher, timeout):
-            supported_ciphers.append(cipher)
-
-    weak_ciphers = ["LOW", "EXP", "eNULL", "aNULL", "RC4", "DES", "MD5", "DH", "ADH"]
-    for cipher in supported_ciphers:
-        if cipher in weak_ciphers:
-            return True
-
-    return False
