+++ src/nettacker/modules/scan/ssl_signed_certificate.yaml
@@ -1,10 +1,9 @@
 info:
-  name: ssl_version_scan
+  name: ssl_certificate_scan
   author: Captain-T2004
   severity: 6
-  description: check if ssl version is unsafe or uses any bad ciphers.
+  description: check if there are any ssl_certificate vulnerabilities present 
   reference:
-    # Need to change this
     - https://www.beyondsecurity.com/resources/vulnerabilities/ssl-certificate-expiry
   profiles:
     - scan
@@ -13,7 +12,7 @@ info:
 payloads:
   - library: ssl
     steps:
-      - method: ssl_version_and_cipher_scan
+      - method: ssl_certificate_scan
         timeout: 3
         host: "{target}"
         ports:
@@ -29,10 +28,12 @@ payloads:
         response:
           condition_type: or
           conditions:
+            self_signed:
+              reverse: false
             grouped_conditions:
               condition_type: and
               conditions:
-                weak_version:
+                weak_signing_algo:
                   reverse: false
-                ssl_version:
+                signing_algo:
                   reverse: false
